% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_fill_nopar.R
\name{code_fill_nopar}
\alias{code_fill_nopar}
\title{Fill Out Level 1 or 2 Habitat Codes}
\usage{
code_fill_nopar(x, level1.recode, level2.recode, subset, verbose)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}

\item{level1.recode}{(optional) Logical. If Level 2 habitat info exists, should the corresponding Level 1 habitat be given the same info? See Details. Default is TRUE.}

\item{level2.recode}{(optional) Logical. If Level 1 habitat info exists, should the corresponding Level 2 habitats be given the same info? See Details. Default is TRUE.}

\item{subset}{(optional). Numeric. Specify proportion of x to run for between 0 and 1. Defaults to 1.}

\item{verbose}{(optional) Numeric. If 0 gives no progress update, if 1 prints progress bar, if 2 prints 1 row per task completed. Default is 1.}
}
\value{
A dataframe in wide format (one column per habitat type and one row per species' season).
}
\description{
Optionally fill out Level 2 habitat codes if Level 1 codes exist and vice versa.
}
\details{
The \code{level1.recode} parameter will fill upstream Level 1 habitat categories if any Level 2 categories have habitat information. e.g. if habitat category 2.1 (Dry Savanna) is suitable and of major importance (coded as 31), habitat category 2 (Savanna) will be filled with this info too.

Similarly, \code{level2.recode} will fill all downstream Level 2 habitat categories if the corresponding Level 1 category is filled. e.g. if if habitat category 2 (Savanna) is suitable and of major importance (coded as 31), habitat categories 2.1 (Dry Savanna) and 2.2 (Moist Savanna) will be given the same information.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
\keyword{internal}
