% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_RL_fetch.R
\docType{data}
\name{example_RL_fetch}
\alias{example_RL_fetch}
\title{RL_fetch() example}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(example_RL_fetch)
}
\description{
Example data running RL_fetch() for "Panthera leo", "Parus major", "Pinus oocarpa".
}
\keyword{datasets}
\keyword{internal}
