% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nopar.R
\name{fetch_nopar}
\alias{fetch_nopar}
\title{Fetch Red List data (non-parallel)}
\usage{
fetch_nopar(x, key, query, subset, sleep_dur, verbose)
}
\arguments{
\item{x}{Vector of species names or IDs.}

\item{key}{Character string. IUCN Red List API key - available from https://apiv3.iucnredlist.org/api/v3/token}

\item{query}{Character string. 'name' if querying by species name; 'ID' if querying by IUCN Red List ID.}

\item{subset}{(optional). Numeric. Specify proportion of x to run for between 0 and 1. Defaults to 1.}

\item{sleep_dur}{(optional) Numeric. Duration of sleep between API calls in seconds. Defaults to 2 seconds.}

\item{verbose}{(optional) Numeric. If 0 gives no progress update, if 1 prints progress bar, if 2 prints 1 row per task completed. Default is 1.}
}
\value{
A dataframe in wide format (one column per habitat type and one row per species' season).
}
\description{
Download Red List habitat and elevation data from the Red List API.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
\keyword{internal}
