% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hab_col_positions.R
\name{hab_col_positions}
\alias{hab_col_positions}
\title{Level 1 and Level2 habitat column positions}
\usage{
hab_col_positions(x = NatureMapRedList::wideform())
}
\arguments{
\item{x}{Dataframe with column names in the form iucn_n or iucn_n_n where n is an integer habitat code}
}
\value{
A list of column positions for level 1 and level 2 habitat categories
}
\description{
Work out positions of level 1 and 2 columns from wide form data frame
}
\keyword{internal}
