% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hab_conversion_lut.R
\docType{data}
\name{hab_conversion_lut}
\alias{hab_conversion_lut}
\title{IUCN Red List Habitat Category New Name Look-up Table}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(hab_conversion_lut)
}
\description{
Look-up table for converting IUCN Habitat categories to new integer names.
}
\keyword{datasets}
