% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats.R
\docType{data}
\name{habitats}
\alias{habitats}
\title{IUCN Red List Habitat Categories}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://www.iucnredlist.org/resources/habitat-classification-scheme}{IUCN Red List Habitat Classification Scheme}
}
\usage{
data(habitats)
}
\description{
Data from the IUCN Red List's Habitat Classification (Version 3.1) using the December 2012 draft working document.
}
\keyword{datasets}
