% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/major_importance.R
\docType{data}
\name{major_importance}
\alias{major_importance}
\title{Nature Map Red List Habitat Major Importance Codes}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(major_importance)
}
\description{
Codes for habitat major importance Somewhat arbitrarily created. Alter this data to alter functions.
}
\keyword{datasets}
