% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packTest.R
\name{packTest}
\alias{packTest}
\title{Check necessary packages are loaded}
\usage{
packTest(pack)
}
\arguments{
\item{pack}{Character. Package name.}
}
\description{
Internal Function. Checks packages necessary for operation are loaded. Stops execution if not
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
\keyword{internal}
