% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_code_fill.R
\name{RL_code_fill}
\alias{RL_code_fill}
\title{Fill Out Level 1 or 2 Habitat Codes}
\usage{
RL_code_fill(
  x,
  level1.recode = T,
  level2.recode = T,
  subset = 1,
  parallel = T,
  num.cores = parallel::detectCores() - 1,
  verbose = 1
)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}

\item{level1.recode}{(optional) Logical. If Level 2 habitat info exists, should the corresponding Level 1 habitat be given the same info? See Details. Default is TRUE.}

\item{level2.recode}{(optional) Logical. If Level 1 habitat info exists, should the corresponding Level 2 habitats be given the same info? See Details. Default is TRUE.}

\item{subset}{(optional). Numeric. Specify proportion of x to run for between 0 and 1. Defaults to 1.}

\item{parallel}{(optional) Logical. If TRUE uses multiple cores. Default is TRUE.}

\item{num.cores}{(optional) Numeric. Specify number of cores to use if running in parallel. Default is number of CPU cores available - 1.}

\item{verbose}{(optional) Numeric. If 0 gives no progress update, if 1 prints progress bar, if 2 prints 1 row per task completed. Default is 1.}
}
\value{
A dataframe in wide format (one column per habitat type and one row per species' season).
}
\description{
Optionally fill out Level 2 habitat codes if Level 1 codes exist and vice versa.
}
\details{
The \code{level1.recode} parameter will fill upstream Level 1 habitat categories if any Level 2 categories have habitat information. e.g. if habitat category 2.1 (Dry Savanna) is suitable and of major importance (coded as 31), habitat category 2 (Savanna) will be filled with this info too.

Similarly, \code{level2.recode} will fill all downstream Level 2 habitat categories if the corresponding Level 1 category is filled. e.g. if if habitat category 2 (Savanna) is suitable and of major importance (coded as 31), habitat categories 2.1 (Dry Savanna) and 2.2 (Moist Savanna) will be given the same information.

NOTE: Running this function multiple times will fill out Level 1/Level 2 habitats which only exist because of a previous function call. It is advisable only to run this function once.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
