% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_col_rename.R
\name{RL_col_rename}
\alias{RL_col_rename}
\title{Rename Habitat Columns to New Names (GEE use)}
\usage{
RL_col_rename(x)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Rename columns to new names - e.g. iucn_1_1 becomes 110.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
