% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_crosswalk.R
\name{RL_crosswalk}
\alias{RL_crosswalk}
\title{Crosswalk habitat classes to a look-up table of other values}
\usage{
RL_crosswalk(
  x,
  lut = redlistManipulatr::hab_conversion_lut,
  new.code.colname = NULL,
  old.code.colname = NULL,
  recode = "no"
)
}
\arguments{
\item{x}{A wide-format \code{dataframe} with one column per habitat category. As output by \code{RL_fetch()}.}

\item{lut}{(optional) A \code{dataframe} look-up table of values with two columns, one for new values, one for Red List habitat classes. This can be one-to-one, one-to-many, or many-to-many. If many-to-many, the highest suitability-major importance combination is selected as the preferred value for the new value - see \code{suitability_ordered}. See details for default.}

\item{new.code.colname}{(optional) Character. What is the name of the column of \code{lut} with the new values in? Defaults to being the first column.}

\item{old.code.colname}{(optional) Character. What is the name of the column of \code{lut} with the old (Red List habitat class) values in? Defaults to being the second column.}

\item{recode}{(optional) Character. Should values be recoded? "no" gives the most favourable suitability-major importance combinations, "one" codes all habitats with suitability-major importance values as 1, "colname" codes all habitats with suitability-major importance values as the column name value (the same as the new values in lut). Defaults to "no".}
}
\value{
A \code{dataframe} in wide format (one column per new value of habitat type).
}
\description{
Rename columns to new names from look-up table - e.g. iucn_1_1 becomes 110. Then optionally replaces suitability/major importance values with the column name or 1. Typically used to convert between habitat classes and land cover classes.
}
\details{
Use should follow that of RL_subset_acceptable() if only some habitat suitabilities/major importances are wanted and cell values are being recoded.

The default for \code{lut} is the \code{hab_conversion_lut} dataset which just converts habitat classes to integer values - e.g. 1.1 -> 110; 9.8.1 -> 981.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
