% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_GEE_ready.R
\name{RL_GEE_ready}
\alias{RL_GEE_ready}
\title{Change wide form habitat column names and give cells their value (for GEE)}
\usage{
RL_GEE_ready(x)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Rename columns to new names - e.g. iucn_1_1 becomes 110. Then replaces suitability/major importance values with the column name. Use should follow that of RL_subset_acceptable() to ready for GEE output.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
