% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_reformat_long.R
\name{RL_reformat_long}
\alias{RL_reformat_long}
\title{Reformat wide format data to long format}
\usage{
RL_reformat_long(x, na.rm = TRUE)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch() or RL_code_fill().}

\item{na.rm}{(optional). Logical. Remove NA rows to compress output. See Details. Defaults to TRUE.}
}
\value{
A dataframe in long format (one row per habitat type).
}
\description{
Reformat wide format data (they way RL_fetch() retrieves data) to long format.
}
\details{
If \code{na.rm} is \code{TRUE} (the default) then rows lacking habitat information will be removed. On pivoting, all habitat columns are retained - even those with no information - meaning that the output has many NA rows where a species is not found at all in a habitat (e.g. polar bears in Hot Desert). This is normally undesirable, so these rows are removed by default. If you would like to keep these rows then set \code{na.rm} to \code{FALSE}.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
