% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_reformat_wide.R
\name{RL_reformat_wide}
\alias{RL_reformat_wide}
\title{Reformat long format data to wide format}
\usage{
RL_reformat_wide(x)
}
\arguments{
\item{x}{A long format dataframe with one row per habitat category. As output by RL_reformat_long().}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Reformat long format data (output by RL_reformat_long()) to wide format (needed for RL_code_fill() etc).
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
