% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_season_recode.R
\name{RL_season_recode}
\alias{RL_season_recode}
\title{Recode NA seasons to Resident}
\usage{
RL_season_recode(x)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Recodes seasons coded on the Red List as NA (and pulled down by RL_fetch() as 999) to Resident (1).
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
