% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_sp_list.R
\name{RL_sp_list}
\alias{RL_sp_list}
\title{Obtain list of all species on Red List}
\usage{
RL_sp_list(
  key,
  retain.taxonomic.info = FALSE,
  retain.population = FALSE,
  retain.category = TRUE
)
}
\arguments{
\item{key}{Character string. IUCN Red List API key - available from https://apiv3.iucnredlist.org/api/v3/token}

\item{retain.taxonomic.info}{(optional) Character. Options to retain all of the taxonomic info columns ("all"), only Kingdom and Class ("some"), or none ("none"). If "none", species binomial name is still provided. Defaults to "some".}

\item{retain.population}{(optional) Logical. Retain the population information? Defaults to FALSE.}

\item{retain.category}{(optional) Logical. Retain the assessment category? Defaults to TRUE.}
}
\value{
A dataframe with one row per species containing IUCN ID (column taxonid), taxonomic information, and Red List category.
}
\description{
Obtain a list and basic info about all species on the Red List. This is a wrapper function for rredlist::rl_sp() used for all species.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
