% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_subset_acceptable.R
\name{RL_subset_acceptable}
\alias{RL_subset_acceptable}
\title{Subset to acceptable suitabilities and major importances of habitats}
\usage{
RL_subset_acceptable(
  x,
  acceptable_suitabilities = NA,
  acceptable_importances = NA,
  na.rm = TRUE
)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}

\item{acceptable_suitabilities}{(optional) A vector of suitabilities deemed acceptable. Must be in the form \code{c(3,2)}. See details. Defaults to accepting all.}

\item{acceptable_importances}{(optional) A vector of major importances deemed acceptable. Must be in the form \code{c(1,3)}. See details. Defaults to accepting all.}

\item{na.rm}{(optional) Logical. If a row has all habitats recoded to NA, should it be removed? Default is TRUE.}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Subset data to only retain desired suitabilities and major importances. Other suitabilities/major importances are recoded as NA. Rows containing only NA values can then be removed.
}
\details{
See \code{suitability} and \code{major_importance} for code values corresponding to each suitability/major importance name.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
