% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_subset_seasons.R
\name{RL_subset_seasons}
\alias{RL_subset_seasons}
\title{Subset to species-season combinations}
\usage{
RL_subset_seasons(
  x,
  season_df,
  query,
  species.col.name = NA,
  season.col.name = NA,
  fill.missing.seasons = NA,
  retain.na.seasons = FALSE,
  retain.missing.sp = TRUE
)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}

\item{season_df}{A dataframe consisting of a column for species and a column for season. A species with multiple seasons desired should have one row per season. See Details.}

\item{query}{Character. Are species represented by names ("name") or IUCN IDs ("ID")?}

\item{species.col.name}{(optional) Character. What column name of \code{season_df} are the species names/IDs found in. Defaults to the first column of the dataframe.}

\item{season.col.name}{(optional) Character. What column name of \code{season_df} are the seasons found in. Defaults to the second column of the dataframe.}

\item{fill.missing.seasons}{(optional) Vector. If seasons are present in \code{season_df} but not on the Red List, should data be copied from seasons which are present and if so which? Acceptable values are season codes - see \code{seasons}. Defaults to NA (no copying).}

\item{retain.na.seasons}{(optional) Logical. Some seasons are coded as NA on the Red List. RL_fetch() pulls these down as 999. Would you like to retain these seasons? Defaults to FALSE.}

\item{retain.missing.sp}{(optional) Logical. If a species is not found in season_df, should it be retained? Defaults to TRUE.}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Subset wide format data to species-season combinations. Also allows copying of data for missing desired seasons.
}
\details{
\code{season_df} should have seasons coded numerically as per \code{redlistManipulatr::seasons$Code}.

If the data for a season doesn't exist on the RedList (e.g. a bird may have breeding season and non-breeding season data but no data for resident), then \code{fill.missing.seasons} lets you say which seasons you want to copy data from. This can include any of the Red List season codes - for instance \code{fill.missing.seasons = c(1,2,5)} would mean that for a season we're missing data for, we look at the data (if it exists) for the \code{resident}, \code{breeding season}, and \code{seasonal occurrence uncertain} seasons and select the most suitable combination - e.g. if \code{resident} has habitat 1.2 as \code{suitable} \code{major importance Yes} and \code{breeding season} has it as \code{marginal} \code{major importance No} then we take the value from \code{resident} because it has a higher suitability.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
