% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_suit_recode.R
\name{RL_suit_recode}
\alias{RL_suit_recode}
\title{Recode NA suitabilities to Suitable}
\usage{
RL_suit_recode(x)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Recodes habitats coded on the Red List as NA (and pulled down by RL_fetch() as 999) to Suitable (3).
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
