% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_threats.R
\name{RL_threats}
\alias{RL_threats}
\title{Fetch Red List threats data}
\usage{
RL_threats(
  x,
  col.name,
  key,
  query,
  subset = 1,
  sleep_dur = 2,
  parallel = TRUE,
  num.cores = parallel::detectCores() - 1,
  verbose = 1
)
}
\arguments{
\item{x}{List, vector, or dataframe with a column corresponding to binomial species names or IUCN Red List IDs.}

\item{col.name}{Character string. Column name if x is a dataframe. Not needed if x is a list or vector.}

\item{key}{Character string. IUCN Red List API key - available from https://apiv3.iucnredlist.org/api/v3/token}

\item{query}{Character string. 'name' if querying by species name; 'ID' if querying by IUCN Red List ID.}

\item{subset}{(optional). Numeric. Specify proportion of x to run for between 0 and 1. Defaults to 1.}

\item{sleep_dur}{(optional) Numeric. Duration of sleep between API calls in seconds. Defaults to 2 seconds.}

\item{parallel}{(optional) Logical. If TRUE uses multiple cores. Default is TRUE.}

\item{num.cores}{(optional) Numeric. Specify number of cores to use if running in parallel. Default is number of CPU cores available - 1.}

\item{verbose}{(optional) Numeric. If 0 gives no progress update, if 1 prints progress bar, if 2 prints 1 row per task completed. Default is 1.}
}
\value{
A dataframe in long format (one row per species-threat combination).
}
\description{
Download Red List threats data from the Red List API.
}
\details{
A code value of "999" indicates that no information for that species could be found.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
