% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_value_replace.R
\name{RL_value_replace}
\alias{RL_value_replace}
\title{Column names to values}
\usage{
RL_value_replace(x)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Replaces suitability/major importance values with the column name. Use should follow that of RL_col_rename() and RL_subset_acceptable() to ready for GEE output.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
