% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\docType{data}
\name{seasons}
\alias{seasons}
\title{IUCN Red List Seasonal Distribution Codes}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://www.iucnredlist.org/resources/mappingstandards}{Mapping Standards and Data Quality for IUCN Red List Spatial Data}
}
\usage{
data(seasons)
}
\description{
Definitions for the IUCN Red List's Seasonal Distribution Codes. Taken from the 'Mapping Standards and Data Quality for the IUCN Red List: Categories and Criteria.' Version 1.16
}
\keyword{datasets}
