% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability.R
\docType{data}
\name{suitability}
\alias{suitability}
\title{Nature Map Red List Suitability Codes}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(suitability)
}
\description{
Codes for habitat suitability. Somewhat arbitrarily created. Alter this data to alter functions.
}
\keyword{datasets}
