% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_ordered.R
\docType{data}
\name{suitability_ordered}
\alias{suitability_ordered}
\title{NatureMap Red List Suitability/Major Importance Code Hierarchy}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(suitability_ordered)
}
\description{
Hierarchy of suitability-major importance codes (from first to last). Used internally.
}
\keyword{datasets}
