% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threatform.R
\name{threatform}
\alias{threatform}
\title{Long format dataframe for RL_threats}
\usage{
threatform(threats = redlistManipulatr::threats)
}
\arguments{
\item{threats}{Dataframe of Red List threat categories. Taken from redlistManipulatr::threats}
}
\value{
A dataframe in long format (one row per threat-species combination).
}
\description{
Internal Function. Makes a long format dataframe to output RL_threats() into according to the Red List threats.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
\keyword{internal}
