% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threats.R
\docType{data}
\name{threats}
\alias{threats}
\title{IUCN Red List Threat Categories}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://www.iucnredlist.org/resources/threat-classification-scheme}{IUCN Red List Threat Classification Scheme}
}
\usage{
data(threats)
}
\description{
Data from the IUCN Red List's Threat Classification (Version 3.2) using the December 2012 draft working document.
}
\keyword{datasets}
