% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideform.R
\name{wideform}
\alias{wideform}
\title{Wide format dataframe for RL_fetch()}
\usage{
wideform(habs = redlistManipulatr::habitats, seas = redlistManipulatr::seasons)
}
\arguments{
\item{habs}{Dataframe of Red List habitat categories. Taken from redlistManipulatr::habitats.}

\item{seas}{Dataframe of Red List season codes Taken from redlistManipulatr::seasons.}
}
\value{
A dataframe in wide format (one column per habitat type and one row per species).
}
\description{
Internal Function. Makes a wide format dataframe to output RL_fetch() into according to the Red List level2 habitats. The number of rows corresponds to the number of seasons coded by the Red List.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
\keyword{internal}
