% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_elevation_check.R
\name{RL_elevation_check}
\alias{RL_elevation_check}
\title{Checks min/max elevations and recodes if min > max}
\usage{
RL_elevation_check(x)
}
\arguments{
\item{x}{A dataframe in wide or long form, as output by \code{RL_fetch()} and \code{RL_reformat_long()} respectively.}
}
\value{
A dataframe in the same format as the input.
}
\description{
Checks all species' elevations and if min > max recodes min to -99999 and max to 99999
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
