% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_subset_seasons.R
\name{RL_subset_seasons}
\alias{RL_subset_seasons}
\title{Subset to species-season combinations}
\usage{
RL_subset_seasons(
  x,
  season_df,
  query,
  species.col.name = NA,
  season.col.name = NA,
  fill.missing.seasons = NA,
  retain.na.seasons = FALSE,
  retain.missing.sp = TRUE
)
}
\arguments{
\item{x}{A wide format dataframe with one column per habitat category. As output by RL_fetch().}

\item{season_df}{A dataframe consisting of a column for species and a column for season. A species with multiple seasons desired should have one row per season. See Details.}

\item{query}{Character. Are species represented by names ("name") or IUCN IDs ("ID")?}

\item{species.col.name}{(optional) Character. What column name of \code{season_df} are the species names/IDs found in. Defaults to the first column of the dataframe.}

\item{season.col.name}{(optional) Character. What column name of \code{season_df} are the seasons found in. Defaults to the second column of the dataframe.}

\item{fill.missing.seasons}{(optional) Vector. If seasons are present in \code{season_df} but not on the Red List, should data be copied from seasons which are present and if so which? Acceptable values are season codes - see \code{seasons}. Defaults to NA (no copying).}

\item{retain.na.seasons}{(optional) Logical. Some seasons are coded as NA on the Red List. RL_fetch() pulls these down as 999. Would you like to retain these seasons? Defaults to FALSE.}

\item{retain.missing.sp}{(optional) Logical. If a species is not found in season_df, should it be retained? Defaults to TRUE.}
}
\value{
A dataframe in wide format (one column per habitat type).
}
\description{
Subset wide format data to species-season combinations. Also allows copying of data for missing desired seasons
}
\details{
\code{season_df} should have seasons coded numerically as per \code{redlistManipulatr::seasons$Code}.
}
\author{
Matt Lewis, \email{matthewlewis896@gmail.com}
}
